requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
PyMySQL==1.0.2
bcrypt==3.2.2
PyJWT==2.4.0
python-dotenv==0.19.2
requests==2.26.0
google-auth==2.6.0
google-auth-oauthlib==0.5.1
google-api-python-client==2.41.0
stripe==3.5.0
Pillow==8.3.2
pywebpush==1.14.0
cryptography==3.4.8
marshmallow==3.14.1
email-validator==1.1.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
click==7.1.2
gunicorn==20.1.0